api/github/graphql: move packages to follow the standard location
Created with Sturdy
