+++ api/pkg/github/enterprise/graphql/codebase_github_integration_import.go
@@ -1,4 +1,4 @@
-package enterprise
+package graphql
 
 import (
 	"context"
