+++ api/pkg/github/enterprise/graphql/pr/pr_root_resolver.go
@@ -1,4 +1,4 @@
-package enterprise
+package pr
 
 import (
 	"context"
