+++ api/pkg/github/graphql/codebase_github_integration.go
@@ -1,4 +1,4 @@
-package oss
+package graphql
 
 import (
 	"context"
