--- api/pkg/github/graphql/module_enterprise.go
@@ -1,17 +1,0 @@
-//go:build enterprise || cloud
-// +build enterprise cloud
-
-package graphql
-
-import (
-	"getsturdy.com/api/pkg/di"
-	"getsturdy.com/api/pkg/github/graphql/enterprise"
-	"getsturdy.com/api/pkg/graphql/resolvers"
-)
-
-func Module(c *di.Container) {
-	c.Register(enterprise.NewGitHubAccountRootResolver, new(resolvers.GitHubAccountRootResolver))
-	c.Register(enterprise.NewGitHubAppRootResolver)
-	c.Register(enterprise.NewCodebaseGitHubIntegrationRootResolver)
-	c.Register(enterprise.NewGitHubRootResolver)
-}
