+++ api/pkg/github/graphql/pr/module.go
@@ -1,0 +1,7 @@
+package pr
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(NewResolver)
+}
