+++ api/pkg/github/graphql/pr/root.go
@@ -1,4 +1,4 @@
-package oss
+package pr
 
 import (
 	"context"
